Early stop control fixed
